pyproject.toml
[project]
name = "bookstore"
version = "1.0.0"
description = "REST API for Bookstore Application"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "pyjwt[crypto]",
    "tenacity",
    "structlog",
    "httpx",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-sqlalchemy",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-exporter-otlp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "testcontainers[postgres]",
    "ruff",
    "mypy",
    "pre-commit",
    "datamodel-code-generator",
    "sqlacodegen",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bookstore"]

# ---------- Ruff (formatter + linter) ----------

[tool.ruff]
src = ["src", "tests"]
line-length = 130
target-version = "py312"
extend-exclude = ["src/generated"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "S",     # flake8-bandit (security)
    "T20",   # flake8-print
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "S105",  # hardcoded password (false positives in config)
    "S106",  # hardcoded password (false positives in config)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # allow assert in tests

[tool.ruff.lint.isort]
known-first-party = ["bookstore"]

# ---------- Pytest ----------

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: integration tests requiring database",
    "unit: unit tests without external dependencies",
    "security: security-related tests",
]

# ---------- Coverage ----------

[tool.coverage.run]
source = ["bookstore"]
branch = true
omit = [
    "*/alembic/*",
    "*/__init__.py",
    "*/generated/*",
]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

# ---------- Mypy ----------

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["src/generated/"]
